refactor!: better handling of byte mode reading
In addition: Rename top level method of Client GetScriptLines to
GetScriptContent and change the way it produces output. Now it returns
whole script as string, as a consequence of byte reading of scripts
instead of reading line by line. I think the bytes are returned not
without reason and protocol was designed to read it this way.

but... in the end, it doesn't even matter...